requirements.txt
python-dotenv==1.0.1
requests==2.31.0
PyYAML==6.0.1
pytz==2024.1
